+++ Changelog.md
@@ -4,7 +4,7 @@ v0.0.5
 - Asynchronous memory storage.
 - PageManager: drop loaded chunks (on get_open_chunks) and dec chunks_in_cur_page() counter.
 - PageManager: page fragmentation support.
-- Lock-free object pool.
+- Chunk memory pool.
 
 v0.0.4
 ======